pyproject.toml
[tool.poetry]
name = "weather-app"
version = "0.1.0"
description = ""
authors = ["aldanadc <aldanacasal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
xmltodict = "^0.13.0"
pytz = "^2023.3.post1"
cachetools = "^5.3.2"
httpx = "^0.26.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
httpretty = "^1.1.4"
freezegun = "^1.4.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
